requirements.txt
asgiref==3.3.1
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.5
django-cors-headers==3.7.0
django-filter==2.4.0
django-material-admin==1.7.16
django-simple-history==2.12.0
djangorestframework==3.12.2
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
Markdown==3.3.3
MarkupSafe==2.0.1
packaging==21.0
Pillow==8.1.0
psycopg2-binary==2.8.6
pyparsing==2.4.7
pytz==2020.5
requests==2.25.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
